backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
pydantic==1.10.14
aiofiles==23.2.1
